vatavaran/server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
pydantic>=2.8.0
uvicorn>=0.30.0
requests>=2.32.0
pandas>=2.2.0
ipython>=8.18.0
PyYAML>=6.0.0
tiktoken>=0.7.0
openai>=1.54.0
pytz>=2022.7
